+++ layers/tests/test_decorators.py
@@ -6,22 +6,22 @@ from django.test.client import RequestFactory
 from django.test.utils import override_settings
 from django.conf import settings
 
-from layers import reset
+from layers import reset_layer_stacks
+from layers.monkey import apply_monkey
 
 
-BASIC_LAYERS_FROM_SETTINGS = {"tree": ["basic", ["web"]], "current": "basic"}
-WEB_LAYERS_FROM_SETTINGS = {"tree": ["basic", ["web"]], "current": "web"}
-BASIC_LAYERS_FROM_REQUEST = {"tree": ["basic", ["web"]], "test-x-django-layer": "basic"}
-WEB_LAYERS_FROM_REQUEST = {"tree": ["basic", ["web"]], "test-x-django-layer": "web"}
+BASIC_LAYERS_ = {"tree": ["basic", ["web"]], "current": "basic"}
+WEB_LAYERS_ = {"tree": ["basic", ["web"]], "current": "web"}
 
 
 class DecoratorFromSettingsTestCase(TestCase):
 
     def setUp(self):
         super(DecoratorFromSettingsTestCase, self).setUp()
-        reset()
+        reset_layer_stacks()
+        apply_monkey(force=True)
 
-    @override_settings(LAYERS=BASIC_LAYERS_FROM_SETTINGS)
+    @override_settings(LAYERS=BASIC_LAYERS_)
     def test_exclude_from_layers_basic(self):
         url = reverse("normal-view")
         response = self.client.get(url)
@@ -35,7 +35,7 @@ class DecoratorFromSettingsTestCase(TestCase):
         self.assertEqual(response.status_code, 200)
         self.failUnless("is not available for your device" in result)
 
-    @override_settings(LAYERS=WEB_LAYERS_FROM_SETTINGS)
+    @override_settings(LAYERS=WEB_LAYERS_)
     def test_exclude_from_layers_web(self):
         url = reverse("normal-view")
         response = self.client.get(url)
@@ -54,25 +54,24 @@ class DecoratorFromRequestTestCase(TestCase):
 
     def setUp(self):
         super(DecoratorFromRequestTestCase, self).setUp()
-        reset()
+        reset_layer_stacks()
+        apply_monkey()
 
-    @override_settings(LAYERS=BASIC_LAYERS_FROM_REQUEST)
+    @override_settings(LAYERS=BASIC_LAYERS_)
     def test_exclude_from_layers_basic(self):
-        print "FIRST GET"
         url = reverse("normal-view")
         response = self.client.get(url, **{"X-Django-Layer": "basic"})
         result = response.content
         self.assertEqual(response.status_code, 200)
         self.failUnless("This is a normal view" in result)
 
-        print "SECOND GET"
         url = reverse("web-only-view")
         response = self.client.get(url, **{"X-Django-Layer": "basic"})
         result = response.content
         self.assertEqual(response.status_code, 200)
         self.failUnless("is not available for your device" in result)
 
-    @override_settings(LAYERS=WEB_LAYERS_FROM_REQUEST)
+    @override_settings(LAYERS=WEB_LAYERS_)
     def test_exclude_from_layers_web(self):
         url = reverse("normal-view")
         response = self.client.get(url, **{"X-Django-Layer": "web"})