+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Changelog
 
+#### next
+1. Django 1.10 compatibility.
+2. Make it possible to determine the layer from the request. This removes the need for a Django process per layer.
+
 #### 1.9
 1. Drop Django 1.6 compatibility. Django 1.9 is supported and tested.